+++ ks_includes/screen_panel.py
@@ -1,5 +1,6 @@
-import logging
 import datetime
+import logging
+
 import gi
 
 gi.require_version("Gtk", "3.0")
@@ -138,7 +139,7 @@ class ScreenPanel:
         hours = seconds // 3600
         hour_units = ngettext("hour", "hours", hours)
         seconds %= 3600
-        minutes = round(seconds / 60)
+        minutes = seconds // 60
         min_units = ngettext("minute", "minutes", minutes)
         seconds %= 60
         sec_units = ngettext("second", "seconds", seconds)