+++ panels/base_panel.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 import logging
-# import psutil
 import gi
 
 import netifaces
@@ -11,6 +10,13 @@ from jinja2 import Environment
 from datetime import datetime
 from math import log
 from ks_includes.screen_panel import ScreenPanel
+
+try:
+    import psutil
+    psutil_available = True
+except ImportError:
+    psutil_available = False
+    logging.debug("psutil is not installed. Unable to do battery check.")
 
 
 class BasePanel(ScreenPanel):
@@ -410,14 +416,14 @@ class BasePanel(ScreenPanel):
             return self.battery_icons['unknown']
 
     def battery_percentage(self):
-        # battery = psutil.sensors_battery()
-        battery = None
+        if not psutil_available:
+            return False
+        battery = psutil.sensors_battery()
         if battery and battery.percent:
             self.labels['battery_icon'].set_from_pixbuf(
                 self.get_battery_icon(battery.percent, battery.power_plugged)
             )
             self.labels['battery'].set_text(f'{battery.percent:.0f}%')
-            logging.debug(f"Battery: {battery.percent}% Power plugged in: {'Yes' if battery.power_plugged else 'No'}")
             self.control['battery_box'].show()
             return True
         else: