+++ panels/main_menu.py
@@ -4,6 +4,7 @@
 # *********************************
 
 import logging
+
 import gi
 
 gi.require_version("Gtk", "3.0")