+++ panels/menu.py
@@ -1,5 +1,6 @@
-import logging
 import json
+import logging
+
 import gi
 
 gi.require_version("Gtk", "3.0")
@@ -66,6 +67,8 @@ class Panel(ScreenPanel):
 
             if item['panel']:
                 b.connect("clicked", self.menu_item_clicked, item)
+            elif item['method'] == "ks_confirm_save":
+                b.connect("clicked", self._screen.confirm_save)
             elif item['method']:
                 params = {}