+++ panels/network.py
@@ -1,12 +1,13 @@
-import subprocess
 import logging
 import os
+
 import gi
 
 gi.require_version("Gtk", "3.0")
 from gi.repository import Gtk, GLib, Pango
 from ks_includes.screen_panel import ScreenPanel
 from ks_includes.sdbus_nm import SdbusNm
+from datetime import datetime
 
 
 class Panel(ScreenPanel):
@@ -14,6 +15,7 @@ class Panel(ScreenPanel):
     def __init__(self, screen, title):
         title = title or _("Network")
         super().__init__(screen, title)
+        self.last_drop_time = datetime.now()
         self.show_add = False
         try:
             self.sdbus_nm = SdbusNm(self.popup_callback)
@@ -28,7 +30,7 @@ class Panel(ScreenPanel):
             message = (
                 _("Failed to initialize") + "\n"
                 + "This panel needs NetworkManager installed into the system\n"
-                + "And the apropriate permissions, without them it will not function.\n"
+                + "And the appropriate permissions, without them it will not function.\n"
                 + f"\n{e}\n"
             )
             self.error_box.add(
@@ -263,6 +265,18 @@ class Panel(ScreenPanel):
         del self.networks[bssid]
         return
 
+    def on_popup_shown(self, combo_box, params):
+        if combo_box.get_property("popup-shown"):
+            logging.debug("Dropdown popup show")
+            self.last_drop_time = datetime.now()
+        else:
+            elapsed = (datetime.now() - self.last_drop_time).total_seconds()
+            if elapsed < 0.2:
+                logging.debug(f"Dropdown closed too fast ({elapsed}s)")
+                GLib.timeout_add(50, combo_box.popup)
+                return
+            logging.debug("Dropdown popup close")
+
     def show_add_network(self, widget, ssid):
         if self.show_add:
             return
@@ -274,12 +288,14 @@ class Panel(ScreenPanel):
             del self.labels['add_network']
 
         eap_method = Gtk.ComboBoxText(hexpand=True)
+        eap_method.connect("notify::popup-shown", self.on_popup_shown)
         for method in ("peap", "ttls", "pwd", "leap", "md5"):
             eap_method.append(method, method.upper())
         self.labels['network_eap_method'] = eap_method
         eap_method.set_active(0)
 
         phase2 = Gtk.ComboBoxText(hexpand=True)
+        phase2.connect("notify::popup-shown", self.on_popup_shown)
         for method in ("mschapv2", "gtc", "pap", "chap", "mschap", "disabled"):
             phase2.append(method, method.upper())
         self.labels['network_phase2'] = phase2
@@ -290,11 +306,13 @@ class Panel(ScreenPanel):
         auth_selection_box.add(self.labels['network_phase2'])
 
         self.labels['network_identity'] = Gtk.Entry(hexpand=True, no_show_all=True)
-        self.labels['network_identity'].connect("focus-in-event", self._screen.show_keyboard)
+        self.labels['network_identity'].connect("touch-event", self._screen.show_keyboard)
+        self.labels['network_identity'].connect("button-press-event", self._screen.show_keyboard)
 
         self.labels['network_psk'] = Gtk.Entry(hexpand=True)
         self.labels['network_psk'].connect("activate", self.add_new_network, ssid)
-        self.labels['network_psk'].connect("focus-in-event", self._screen.show_keyboard)
+        self.labels['network_psk'].connect("touch-event", self._screen.show_keyboard)
+        self.labels['network_psk'].connect("button-press-event", self._screen.show_keyboard)
 
         save = self._gtk.Button("sd", _("Save"), "color3")
         save.set_hexpand(False)